+++ demo/example.js
@@ -37,10 +37,13 @@ export const modal = `
 
   <Input type="text" name="subject" label="Subject" required />
   <Textarea name="message" label="Message" maxLength="500" />
+
   <ConversationsSelect
     name="shareWith"
     label="Share with..."
     required
+    include="public im"
+    excludeBotUsers
     responseUrlEnabled
   />